news/requirements.txt
Django~=4.1
django-filter~=22.1
django-adminlte-3~=0.1.6
djangorestframework~=3.13.1
mysqlclient~=2.1.1